\name{gp.ext2terminal}
\alias{gp.ext2terminal}
\title{Determine a suitable terminal from a file extension}
\description{\code{gp.ext2terminal} }
\usage{gp.ext2terminal(filetype='PNG')}
\arguments{	\item{filetype}{ file extension}
}
\value{	terminal name}
\seealso{	\code{\link{gp.set.term}}}
\author{Jose' Gama}
\keyword{programming}
