\name{gp.fit.allprogress}
\alias{gp.fit.allprogress}
\title{Get the value of all parameters from a stderr fit file}
\description{\code{gp.fit.allprogress} returns the value of all parameters from a logged stderr fit file}
\usage{gp.fit.allprogress(fitprogressfile, boolScreenOut = TRUE)}
\arguments{	\item{fitprogressfile}{ logged stderr fit file}
	\item{boolScreenOut}{ optional, if TRUE then the output will also be displayed on the screen}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
