\name{gp.fit.progress}
\alias{gp.fit.progress}
\title{Get the value of a parameter from a stderr fit file}
\description{\code{gp.fit.progress} returns the value of a parameter from a logged stderr fit file}
\usage{gp.fit.progress(fitprogressfile, fitparameter)}
\arguments{	\item{fitprogressfile}{ logged stderr fit file}
	\item{fitparameter}{ fit parameter}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
