\name{gp.hexRGB}
\alias{gp.hexRGB}
\title{RGB sequence, from decimal to hexadecimal, ready for gnuplot}
\description{\code{gp.hexRGB} returns an RGB sequence, from decimal to hexadecimal}
\usage{gp.hexRGB(r,g=NA,b=NA)}
\arguments{	\item{r}{ value for red channel}
\item{g}{ value for green channel}
\item{b}{ value for blue channel}
}
\value{	hexadecimal RGB sequence}
\seealso{	\code{\link{gp.image.rgbchange}}}
\author{Jose' Gama}
\keyword{programming}
