\name{gp.init.save.stderr}
\alias{gp.init.save.stderr}
\title{Initialize gnuplot, save stderr to a log file}
\description{\code{gp.init.save.stderr} initializes gnuplot and saves stderr to a log file}
\usage{gp.init.save.stderr(logfile)}
\arguments{	\item{logfile}{ log file name}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
