\name{gp.isWindowOpen}
\alias{gp.isWindowOpen}
\title{Check if an X11 window is opened}
\description{\code{gp.isWindowOpen} returns TRUE if an X11 window is opened}
\usage{gp.isWindowOpen(windowid)}
\arguments{	\item{windowid}{ X11 window id}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
