\name{gp.killpidX11}
\alias{gp.killpidX11}
\title{Kill X11 windows}
\description{\code{gp.killpidX11} kills X11 windows (gnuplot), useful for testing purposes only- experimental, use at your own risk!}
\usage{gp.killpidX11(windowid=0)}
\arguments{	\item{windowid}{ X11 window id}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
