\name{gp.math3d.plot}
\alias{gp.math3d.plot}
\title{3d plots like Mathematica}
\description{\code{gp.math3d.plot} }
\usage{gp.math3d.plot(foo, xrange=c(0,1), yrange=c(0,1))}
\arguments{	\item{foo}{ mathematical expression}
\item{xrange}{ x-axis range}
\item{yrange}{ y-axis range}
}
\value{	none}
\seealso{	\code{\link{gp.colorhistogram}}}
\author{Ben Ruijl and Jose' Gama}
\keyword{programming}
