\name{gp.pausableterm}
\alias{gp.pausableterm}
\title{Determine if the current terminal can be paused in synch with R}
\description{\code{gp.pausableterm} returns TRUE if the current terminal can be paused}
\usage{gp.pausableterm()}
\value{	boolean}
\seealso{	\code{\link{gp.pause}}}
\author{Jose' Gama}
\keyword{programming}
