\name{gp.pauseX}
\alias{gp.pauseX}
\title{Pauses the system (X11)}
\description{\code{gp.pauseX} pauses the system for a number of seconds and then waits for the user to press a key, X11 only}
\usage{gp.pauseX(delaySecs=2)}
\arguments{	\item{delaySecs}{ number of seconds to wait}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
