\name{gp.pidX11}
\alias{gp.pidX11}
\title{Get the pid from an X11 window}
\description{\code{gp.pidX11} returns the pid from an X11 window (gnuplot)}
\usage{gp.pidX11(windowid=0)}
\arguments{	\item{windowid}{ window id (X11)}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
