\name{gp.plot.once}
\alias{gp.plot.once}
\title{Plots arrays in one function call}
\description{ \code{gp.plot.once} plots arrays for the x and y values without the need to define handles}
\usage{gp.plot.once(title, style, label.x, label.y, x, y, n)}
\arguments{ \item{title}{ title for the graphic}
\item{style}{ gnuplot style (lines, points, linespoints, ...)}
\item{label.x}{ label for the x-axis}
\item{label.y}{ label for the y-axis}
\item{x}{ array of doubles for the x-axis}
\item{y}{ array of doubles for the y-axis}
\item{n}{ number of elements of the list}}
\value{	none}
\seealso{ \code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
