\name{gp.plot.slope}
\alias{gp.plot.slope}
\title{Plots a slope}
\description{\code{gp.plot.slope} plots a slope from a list of doubles}
\usage{gp.plot.slope(handle, a, b, title)}
\arguments{ \item{handle}{ handle to the connection}
\item{a}{ list of doubles}
\item{b}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{ \code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
