\name{gp.plot.x}
\alias{gp.plot.x}
\title{Plot one list of doubles in 2D}
\description{\code{gp.plot.x} plots one list of doubles in 2D}
\usage{gp.plot.x(handle, d, n, title)}
\arguments{	\item{handle}{ handle to the connection}
\item{d}{ list of doubles}
\item{n}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
