\name{gp.plot.xy}
\alias{gp.plot.xy}
\title{Plot two list of doubles in 2D}
\description{\code{gp.plot.xy} plots two list of doubles in 2D}
\usage{gp.plot.xy(handle, x, y, n, title)}
\arguments{ \item{handle}{ handle to the connection}
\item{x}{ list of doubles for the x-axis}
\item{y}{ list of doubles for the y-axis}
\item{n}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{ \code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
