\name{gp.PNG2color}
\alias{gp.PNG2color}
\title{Convert PNG to text}
\description{\code{gp.PNG2color} converts a PNG file to a text format readable by gnuplot}
\usage{gp.PNG2color(fileName)}
\arguments{	\item{fileName}{ PNG file name}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
