\name{gp.PNG2RGB}
\alias{gp.PNG2RGB}
\title{Convert PNG to RGB}
\description{\code{gp.PNG2RGB} converts a PNG file to an RGB or RGBA file}
\usage{gp.PNG2RGB(PNGfile, RGBfile,forceRGB=FALSE)}
\arguments{	\item{PNGfile}{ PNG file}
	\item{RGBfile}{ RGB or RGBA file}
	\item{forceRGB}{ if forceRGB is TRUE then the alpha channel is ignored}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
