\name{gp.R2plot}
\alias{gp.R2plot}
\title{Save 2D R to gnuplot}
\description{\code{gp.R2plot} saves the output from a 2D R function to a file readable by gnuplot}
\usage{gp.R2plot(Rfunction,filename,gnuxrange,gnuyrange,gnusamples)}
\arguments{	\item{Rfunction}{ 2D R function}
	\item{filename}{ file name (gnuplot)}
	\item{gnuxrange}{ gnuplot xrange}
	\item{gnuyrange}{ gnuplot yrange}
	\item{gnusamples}{ gnuplot samples}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
