\name{gp.RGB2PNG}
\alias{gp.RGB2PNG}
\title{convert RGB to PNG}
\description{\code{gp.RGB2PNG} converts an RGB file to a PNG file}
\usage{gp.RGB2PNG(RGBfile, PNGfile, width, height)}
\arguments{	\item{RGBfile}{ RGB file name}
	\item{PNGfile}{ PNG file name}
	\item{width}{ width of the image}
	\item{height}{ height of the image}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
