\name{gp.run}
\alias{gp.run}
\title{executes gnuplot code directly}
\description{\code{gp.run} returns the gnuplot version}
\usage{gp.run(cmd, optPause=FALSE)}
\arguments{	\item{cmd}{ String with gnuplot commands}
\item{optPause}{ boolean, TRUE means to pause after the execution}
}
\value{	none}
\seealso{	\code{\link{gp.pause}}}
\author{Jose' Gama}
\keyword{programming}
