\name{gp.set.term}
\alias{gp.set.term}
\title{List or change gnuplot's current terminal}
\description{\code{gp.set.term} sets gnuplot's current terminal or get a list of all available terminals}
\usage{gp.set.term(optionalNewTerminal)}
\arguments{	\item{optionalNewTerminal}{ should be either empty or the name of an available terminal}
}
\value{	the name of gnuplot's current terminal}
\seealso{	\code{\link{gp.showterm}}}
\author{Jose' Gama}
\keyword{programming}
