\name{gp.setvariable}
\alias{gp.setvariable}
\title{Set a system or environment variable}
\description{\code{gp.setvariable} sets a system or environment variable "variablename", with value "variabledata"}
\usage{gp.setvariable(handle,variablename,variabledata)}
\arguments{	\item{handle}{ handle to the connection}
	\item{variablename}{ variable name}
	\item{variabledata}{ data value}}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
