\name{gp.show.datafile.binary.filetypes}
\alias{gp.show.datafile.binary.filetypes}
\title{ List gnuplot's RGB binary types}
\description{\code{gp.show.datafile.binary.filetypes} get gnuplot's RGB binary types as a vector}
\usage{gp.show.datafile.binary.filetypes()}
\value{	vector with the names of gnuplot's RGB binary types}
\seealso{	\code{\link{gp.image.plot}}}
\author{Jose' Gama}
\keyword{programming}
