\name{gp.showterm}
\alias{gp.showterm}
\title{ Get gnuplot's current terminal}
\description{\code{gp.showterm} gets gnuplot's current terminal
\code{gp.set.term} sets gnuplot's current terminal or gets a list of all available terminals
}
\usage{gp.showterm(optionalNewTerminal)
}
\arguments{	\item{optionalNewTerminal}{ current terminal}
}
\value{	the name of gnuplot's current terminal
a list of all available terminals
}
\seealso{	\code{\link{gp.ext2terminal}}}
\author{Jose' Gama}
\keyword{programming}
