\name{gp.write.x.csv}
\alias{gp.write.x.csv}
\title{Write single column CSV file}
\description{\code{gp.write.x.csv} writes a single column CSV file for use with gnuplot commands later}
\usage{gp.write.x.csv(fileName,x,n,title)}
\arguments{	\item{fileName}{ CSV file name}
	\item{x}{ A column buffer}
	\item{n}{ number of columns}
	\item{title}{ 	Title to write for the first line of the .csv file, will be preceeded by "#" }}
\value{	none}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\keyword{programming}
