\name{gp.X11.present}
\alias{gp.X11.present}
\title{Check for X11 in the system}
\description{\code{gp.X11.present} returns TRUE if X11 is present in the system}
\usage{gp.X11.present()}
\value{	TRUE if X11 is present in the system}
\seealso{	\code{\link{gp.init}}}
\author{Jose' Gama}
\examples{
\dontrun{gp.X11.present()}
}
\keyword{programming}
