\name{Gpcolorhistogram}
\alias{Gpcolorhistogram}
\title{Plot a color histogram}
\description{\code{Gpcolorhistogram} plots a color histogram}
\usage{Gpcolorhistogram(fileIN, Ncols=1, vectCols=NULL, gtitle='',gxlabel='',gylabel='')}
\arguments{	\item{fileIN}{ input file}
\item{Ncols}{ number of columns to plot}
\item{vectCols}{ column indices to use}
\item{gtitle}{ title}
\item{gxlabel}{ x-axis label}
\item{gylabel}{ y-axis label}
}
\value{	none}
\author{Jose' Gama}
\keyword{programming}
