\name{GpdivergingColormap}
\alias{GpdivergingColormap}
\title{Create diverging colormaps}
\description{\code{GpdivergingColormap} }
\usage{GpdivergingColormap(s,rgb1,rgb2, outColorspace='sRGB')}
\arguments{	\item{s}{ vector with values between zero and one}
\item{rgb1}{ colorspace element for lower bound}
\item{rgb2}{ colorspace element for upper bound}
\item{outColorspace}{ name of the output color space (RGB, sRGB, HLS, HSV, LAB, LUV, PolarLAB, PolarLUV, XYZ)}
}
\value{	none}
\seealso{	\code{\link{Gprun}}}
\author{Kenneth Moreland, Andy Stein and Jose' Gama}
\keyword{programming}
