\name{Gpext2terminal}
\alias{Gpext2terminal}
\title{Determine a suitable terminal from a file extension}
\description{\code{Gpext2terminal} }
\usage{Gpext2terminal(filetype='PNG')}
\arguments{	\item{filetype}{ file extension}
}
\value{	terminal name}
\seealso{	\code{\link{GpsetTerm}}}
\author{Jose' Gama}
\keyword{programming}
