\name{GpfitAllprogress}
\alias{GpfitAllprogress}
\title{Get the value of all parameters from a stderr fit file}
\description{\code{GpfitAllprogress} returns the value of all parameters from a logged stderr fit file}
\usage{GpfitAllprogress(fitprogressfile, boolScreenOut = TRUE)}
\arguments{	\item{fitprogressfile}{ logged stderr fit file}
	\item{boolScreenOut}{ optional, if TRUE then the output will also be displayed on the screen}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
