\name{GpfitProgress}
\alias{GpfitProgress}
\title{Get the value of a parameter from a stderr fit file}
\description{\code{GpfitProgress} returns the value of a parameter from a logged stderr fit file}
\usage{GpfitProgress(fitprogressfile, fitparameter)}
\arguments{	\item{fitprogressfile}{ logged stderr fit file}
	\item{fitparameter}{ fit parameter}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
