\name{GphexRGB}
\alias{GphexRGB}
\title{RGB sequence, from decimal to hexadecimal, ready for gnuplot}
\description{\code{GphexRGB} returns an RGB sequence, from decimal to hexadecimal}
\usage{GphexRGB(r,g=NA,b=NA)}
\arguments{	\item{r}{ value for red channel}
\item{g}{ value for green channel}
\item{b}{ value for blue channel}
}
\value{	hexadecimal RGB sequence}
\seealso{	\code{\link{GpimageRGBchange}}}
\author{Jose' Gama}
\keyword{programming}
