\name{GpinitSaveStderr}
\alias{GpinitSaveStderr}
\title{Initialize gnuplot, save stderr to a log file}
\description{\code{GpinitSaveStderr} initializes gnuplot and saves stderr to a log file}
\usage{GpinitSaveStderr(logfile)}
\arguments{	\item{logfile}{ log file name}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
