\name{GpisWindowOpen}
\alias{GpisWindowOpen}
\title{Check if an X11 window is opened}
\description{\code{GpisWindowOpen} returns TRUE if an X11 window is opened}
\usage{GpisWindowOpen(windowid)}
\arguments{	\item{windowid}{ X11 window id}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
