\name{GpkillpidX11}
\alias{GpkillpidX11}
\title{Kill X11 windows}
\description{\code{GpkillpidX11} kills X11 windows (gnuplot), useful for testing purposes only- experimental, use at your own risk!}
\usage{GpkillpidX11(windowid=0)}
\arguments{	\item{windowid}{ X11 window id}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
