\name{Gpmath3dPlot}
\alias{Gpmath3dPlot}
\title{3d plots like Mathematica}
\description{\code{Gpmath3dPlot} }
\usage{Gpmath3dPlot(foo, xrange=c(0,1), yrange=c(0,1))}
\arguments{	\item{foo}{ mathematical expression}
\item{xrange}{ x-axis range}
\item{yrange}{ y-axis range}
}
\value{	none}
\seealso{	\code{\link{Gpcolorhistogram}}}
\author{Ben Ruijl and Jose' Gama}
\keyword{programming}
