\name{Gppausableterm}
\alias{Gppausableterm}
\title{Determine if the current terminal can be paused in synch with R}
\description{\code{Gppausableterm} returns TRUE if the current terminal can be paused}
\usage{Gppausableterm()}
\value{	boolean}
\seealso{	\code{\link{Gppause}}}
\author{Jose' Gama}
\keyword{programming}
