\name{GppauseX}
\alias{GppauseX}
\title{Pauses the system (X11)}
\description{\code{GppauseX} pauses the system for a number of seconds and then waits for the user to press a key, X11 only}
\usage{GppauseX(delaySecs=2)}
\arguments{	\item{delaySecs}{ number of seconds to wait}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
