\name{GppidX11}
\alias{GppidX11}
\title{Get the pid from an X11 window}
\description{\code{GppidX11} returns the pid from an X11 window (gnuplot)}
\usage{GppidX11(windowid=0)}
\arguments{	\item{windowid}{ window id (X11)}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
