\name{GpplotSlope}
\alias{GpplotSlope}
\title{Plots a slope}
\description{\code{GpplotSlope} plots a slope from a list of doubles}
\usage{GpplotSlope(handle, a, b, title)}
\arguments{ \item{handle}{ handle to the connection}
\item{a}{ list of doubles}
\item{b}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{ \code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
