\name{GpplotX}
\alias{GpplotX}
\title{Plot one list of doubles in 2D}
\description{\code{GpplotX} plots one list of doubles in 2D}
\usage{GpplotX(handle, d, n, title)}
\arguments{	\item{handle}{ handle to the connection}
\item{d}{ list of doubles}
\item{n}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
