\name{GpPNG2color}
\alias{GpPNG2color}
\title{Convert PNG to text}
\description{\code{GpPNG2color} converts a PNG file to a text format readable by gnuplot}
\usage{GpPNG2color(fileName)}
\arguments{	\item{fileName}{ PNG file name}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
