\name{GpPNG2RGB}
\alias{GpPNG2RGB}
\title{Convert PNG to RGB}
\description{\code{GpPNG2RGB} converts a PNG file to an RGB or RGBA file}
\usage{GpPNG2RGB(PNGfile, RGBfile,forceRGB=FALSE)}
\arguments{	\item{PNGfile}{ PNG file}
	\item{RGBfile}{ RGB or RGBA file}
	\item{forceRGB}{ if forceRGB is TRUE then the alpha channel is ignored}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
