\name{GpPNG4DEM}
\alias{GpPNG4DEM}
\title{Convert PNG plus DEM to text}
\description{\code{GpPNG4DEM} converts a PNG file and a DEM tab separated to a text format readable by gnuplot }
\usage{GpPNG4DEM(filePNG, fileDEMtab, file3Ddat)}
\arguments{	\item{filePNG}{ PNG file}
	\item{fileDEMtab}{ DEM tab separated file}
	\item{file3Ddat}{ text file (gnuplot)}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
