\name{Gprun}
\alias{Gprun}
\title{executes gnuplot code directly}
\description{\code{Gprun} returns the gnuplot version}
\usage{Gprun(cmd, optPause=FALSE)}
\arguments{	\item{cmd}{ String with gnuplot commands}
\item{optPause}{ boolean, TRUE means to pause after the execution}
}
\value{	none}
\seealso{	\code{\link{Gppause}}}
\author{Jose' Gama}
\keyword{programming}
