\name{GpsetTerm}
\alias{GpsetTerm}
\title{List or change gnuplot's current terminal}
\description{\code{GpsetTerm} sets gnuplot's current terminal or get a list of all available terminals}
\usage{GpsetTerm(optionalNewTerminal)}
\arguments{	\item{optionalNewTerminal}{ should be either empty or the name of an available terminal}
}
\value{	the name of gnuplot's current terminal}
\seealso{	\code{\link{Gpshowterm}}}
\author{Jose' Gama}
\keyword{programming}
