\name{Gpsetvariable}
\alias{Gpsetvariable}
\title{Set a system or environment variable}
\description{\code{Gpsetvariable} sets a system or environment variable "variablename", with value "variabledata"}
\usage{Gpsetvariable(handle,variablename,variabledata)}
\arguments{	\item{handle}{ handle to the connection}
	\item{variablename}{ variable name}
	\item{variabledata}{ data value}}
\value{	none}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
