\name{GpshowDatafileBinaryFiletypes}
\alias{GpshowDatafileBinaryFiletypes}
\title{ List gnuplot's RGB binary types}
\description{\code{GpshowDatafileBinaryFiletypes} get gnuplot's RGB binary types as a vector}
\usage{GpshowDatafileBinaryFiletypes()}
\value{	vector with the names of gnuplot's RGB binary types}
\seealso{	\code{\link{GpimagePlot}}}
\author{Jose' Gama}
\keyword{programming}
