\name{GpX11Present}
\alias{GpX11Present}
\title{Check for X11 in the system}
\description{\code{GpX11Present} returns TRUE if X11 is present in the system}
\usage{GpX11Present()}
\value{	TRUE if X11 is present in the system}
\seealso{	\code{\link{Gpinit}}}
\author{Jose' Gama}
\examples{
\dontrun{GpX11Present()}
}
\keyword{programming}
