\name{GpplotXY}
\alias{GpplotXY}
\title{Plot two list of doubles in 2D}
\description{\code{GpplotXY} plots two list of doubles in 2D}
\usage{GpplotXY(handle, x, y, n, title)}
\arguments{ \item{handle}{ handle to the connection}
\item{x}{ list of doubles for the x-axis}
\item{y}{ list of doubles for the y-axis}
\item{n}{ number of elements of the list}
\item{title}{ title for the graphic}}
\value{	none}
\seealso{ \code{\link{Gpinit}}}
\author{Jose' Gama}
\keyword{programming}
